serverless-pipeline/src/functions/backup_verifier/requirements.txt
google-cloud-storage==2.*
google-cloud-pubsub==2.*
functions-framework==3.* 